backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pandas
python-multipart
passlib
python-jose